requirements.txt
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
